Add Travis CI config for automatic builds and static analysis
Adding a .travis.yml file with config for automated builds enables
travis-ci to perform some basic continuous integration.  This does not
replace our existing build and test infrastructure, since that tests
many more combinations and platforms.

When added to the repository, this enables automated builds and (basic)
tests for all branches this commit is applied to, and enables coverity
analysis for the coverity_scan branch (since we can only do a limited
number of coverity scans per week with our FOSS account).

Signed-off-by: Steffan Karger <steffan@karger.me>